+++ configurer/linux/alpine.go
@@ -9,13 +9,32 @@ import (
 	"github.com/k0sproject/rig/os/registry"
 )
 
-// RigLinux is a reference to rig's linux struct, renamed to not overlap with configurer.Linux
-type RigLinux os.Linux
+// BaseLinux for tricking go interfaces
+type BaseLinux struct {
+	configurer.Linux
+}
+
+var kubectlInstallScript = []string{
+	`curl -LO "https://dl.k8s.io/release/$(curl -L -s https://dl.k8s.io/release/stable.txt)/bin/linux/amd64/kubectl"`,
+	`curl -LO "https://dl.k8s.io/$(curl -L -s https://dl.k8s.io/release/stable.txt)/bin/linux/amd64/kubectl.sha256"`,
+	`echo "$(<kubectl.sha256) kubectl" | sha256sum --check`,
+	`sudo install -o root -g root -m 0755 kubectl /usr/local/bin/kubectl`,
+}
+
+// InstallKubectl installs kubectl using the curl method
+func (l BaseLinux) InstallKubectl(h os.Host) error {
+	for _, c := range kubectlInstallScript {
+		if err := h.Exec(c); err != nil {
+			return err
+		}
+	}
+	return nil
+}
 
 // Alpine provides OS support for Alpine Linux
 type Alpine struct {
-	RigLinux
-	configurer.Linux
+	os.Linux
+	BaseLinux
 }
 
 func init() {
