+++ configurer/linux/enterpriselinux.go
@@ -1,6 +1,7 @@
 package linux
 
 import (
+	"github.com/k0sproject/k0sctl/configurer"
 	"github.com/k0sproject/rig/os"
 	"github.com/k0sproject/rig/os/linux"
 )
@@ -8,6 +9,7 @@ import (
 // EnterpriseLinux is a base package for several RHEL-like enterprise linux distributions
 type EnterpriseLinux struct {
 	linux.EnterpriseLinux
+	configurer.Linux
 }
 
 // InstallKubectl installs kubectl using the gcloud kubernetes repo
