+++ configurer/linux/sles.go
@@ -1,7 +1,6 @@
 package linux
 
 import (
-	"github.com/k0sproject/k0sctl/configurer"
 	"github.com/k0sproject/rig"
 	"github.com/k0sproject/rig/os"
 	"github.com/k0sproject/rig/os/linux"
@@ -11,7 +10,8 @@ import (
 // SLES provides OS support for Suse SUSE Linux Enterprise Server
 type SLES struct {
 	linux.SLES
-	configurer.Linux
+	os.Linux
+	BaseLinux
 }
 
 func init() {
@@ -23,21 +23,4 @@ func init() {
 			return SLES{}
 		},
 	)
-}
-
-var kubectlInstallScript = []string{
-	`curl -LO "https://dl.k8s.io/release/$(curl -L -s https://dl.k8s.io/release/stable.txt)/bin/linux/amd64/kubectl"`,
-	`curl -LO "https://dl.k8s.io/$(curl -L -s https://dl.k8s.io/release/stable.txt)/bin/linux/amd64/kubectl.sha256"`,
-	`echo "$(<kubectl.sha256) kubectl" | sha256sum --check`,
-	`sudo install -o root -g root -m 0755 kubectl /usr/local/bin/kubectl`,
-}
-
-// InstallKubectl installs kubectl using the curl method
-func (l SLES) InstallKubectl(h os.Host) error {
-	for _, c := range kubectlInstallScript {
-		if err := h.Exec(c); err != nil {
-			return err
-		}
-	}
-	return nil
 }
